Support generating .deb files for the test cases
The tests need to operate on .deb files, but checking those all into the
repository isn't ideal. Instead, we just save the .control files inside
the repository and then generate the .debs dynamically.

Included is a download_files.sh script that can be used to easily
download .changes and .control files.