obs2aptly: Check for & reuse existing packages in the pool
When uploading a new package, there may be another one with the same
name already in the shared pool. Depending on the circumstances, we then
want to either give a hard error or simply reuse the existing one.

In order to keep the tests working with the new SyncAction fields, this
also introduces a new variant, ExpectedSyncAction, for use in the tests.
This lets us change the structure of SyncActions more freely while
letting the tests use a more convenient-to-write structure.

As a side effect of this change, this also seems to work around an
occasional issue related to being unable to overwrite a published
package with a different version of the file. Because the files are now
reused, it's not generally possible to actually have two different
copies of the same package anymore.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>