gitignore: Only ignore 'target' in the repo root
The only Rust-generated 'target' directory is in the root. Ignoring all
others can lead to strange issues such as, for instance, being unable to
find a reference inside a file because that file happened to be saved in
a subdirectory named 'target', and the search tool ignores files in
the gitignore. (Yes, this did happen to me.)

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>