docker: Speed up by using a bind mount for the sources
When doing a build with `cargo` in debug mode we end up with roughly
730MB under `target/` due to dependencies and build artifacts.

Trying to build the Dockerfile after that will `ADD` the whole
workspace including all the previous build artifacts, which takes
a while.

We do not really need to actually capture the sources in the build
container, so we can instead bind mount them and avoid any copying.

The downsides are:

1. building the container twice with no changes to the workspace
   will start the build from scratch, while with `ADD`/`COPY` it
   gets cached
2. `podman` users with SELinux enabled will need to pass
   `--security-opt label=disable`

The caching is really fragile already, since *any* change to the
workspace would invalidate it anyway.

And I assume that Podman users are rather used to passing
`--security-opt label=disable` every time they get permission errors.

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>